# Expose lto and llvm folder compilation flags