--- a/Makefile
+++ b/Makefile
@@ -213,6 +213,9 @@ ifeq (${USE_MERGED_VCPKG_MANIFEST}, 1)
 	CMAKE_VARS:=${CMAKE_VARS} -DVCPKG_MANIFEST_DIR='${PROJ_DIR}build/extension_configuration'
 endif
 
+ifneq ("${LTO}", "")
+	CMAKE_VARS:=${CMAKE_VARS} -DCMAKE_LTO='${LTO}'
+endif
 
 clean:
 	rm -rf build