--- a/scripts/regression_test_runner.py
+++ b/scripts/regression_test_runner.py
@@ -6,9 +6,11 @@ import csv
 import statistics
 import math
 
+
 # Geometric mean of an array of numbers
 def geomean(xs):
     return math.exp(math.fsum(math.log(float(x)) for x in xs) / len(xs))
+
 
 # how many times we will run the experiment, to be sure of the regression
 number_repetitions = 5
@@ -48,10 +50,8 @@ if not os.path.isfile(new_runner):
     print(f"Failed to find new runner {new_runner}")
     exit(1)
 
-complete_timings = {
-    old_runner : [],
-    new_runner : []
-}
+complete_timings = {old_runner: [], new_runner: []}
+
 
 def run_benchmark(runner, benchmark):
     benchmark_args = [runner, benchmark]