This PR expose 2 independent flags to make and cmake invocations that enable link time optimisations and allow relying on an specific LLVM binary folder.
Examples:
```bash
LTO=full make                                          // default options plus passing -flto=full to the compiler/linker
LTO=thin make                                          // default options plus passing -flto=thin to the compiler/linker
CMAKE_LLVM_PATH='~/llvm-project/build' make            // default options using clang++ / clang and llvm-ranlib found in the given folder
LTO=full CMAKE_LLVM_PATH='~/llvm-project/build' make   // default options using clang++ / clang and llvm-ranlib found in the given folder PLUS -flto=full
```

Option are composable, between each other and with other options, but LTO requires the underlying compiler to supports it.
`clang` supports both thin and full options, `gcc` only the full option, while providing a non-supported option will result in a compiler failure.

This PR do not adds either LTO or updated clang to any workflow but for self tests executed in NightlyTests.yml.
Eventually enabling this for distributed binaries is moved to a follow-up PR.

LTO basics
---
LTO (Link Time Optimisation) is basically trading slower compilation times for somewhat more optimised binary produced.
clang exposes also a rebuild friendly ThinLTO (http://blog.llvm.org/2016/06/thinlto-scalable-and-incremental-lto.html) that aims to be possible to turn in also during development.

This are full rebuild times without LTO, LTO=thin and LTO=full on a Mac M2 using the default clang 14:
```
GEN=ninja make			489.60s user 30.12s system 678% cpu 1:16.57 total
GEN=ninja LTO=thin make		900.59s user 40.24s system 663% cpu 2:21.72 total
GEN=ninja LTO=full make		711.20s user 46.07s system 341% cpu 3:41.52 total
```
I tried to estimate performance gains using our current benchmark suite, and speed up for full LTO seems to be, (very unscientifically) about 3% (on geometrical mean of all tests).
More testing / benchmarking is probably required to put a serious number into that. But I'd say the fact that there seems to be no regressions AND some workload are optimised significantly (up to 70% faster) it would make sense to consider this for inclusion.

Idea of this PR is allow, now or in the future, to experiment with this easily.

CMAKE_LLVM_PATH
---
Currently on most workflows we build DuckDB using the default system compiler, that for clang is version 14 both on macos and on ubuntu 22.04.
Current development is at clang 17 (used by duckdb-wasm), while clang 16 is already stable and packaged for example via `brew install llvm`.
On my machine (Mac M2):
```bash
brew install llvm
locate llvm-ranlib
--- /opt/homebrew/Cellar/llvm/16.0.6/bin/llvm-ranlib
CMAKE_LLVM_PATH='/opt/homebrew/Cellar/llvm/16.0.6' make
```
allows to build DuckDB using clang 16.0.6 instead of the stock clang 14.0.3.
More recent compilers allows more optimisation opportunities to be leveraged, also here benchmarking has been done not very seriously but seems to point towards something like 5% improvement in execution speed.

I have done this only for LLVM/clang (since it required llvm-ranlib to be specified), but potentially it could be worth exploring performance gains for more recent gcc versions.

How to roll this in
---
My idea is that having there options more easily available allows more experimenting with this, potentially moving some nightly binaries to use LTO and more recent compilers version and collect feedback to be able to decide whether this is worthy to be turned on also for proper releases.
But input is very welcome, and if someone wants to take this over and give a critical look, you are very welcome!

Note on benchmarking
---
Benchmarking is hard, especially if you do that while trying to prove a point.
I added to the regression test runner a summary such as "new is roughly X% faster | old is roughly Y% faster | about the same". This is done comparing geometrical means. It's a very blunt simplification, do take this with quite some distance, if it does more harm than good it should be removed.

I also added 3 jobs to be executed on nightly that re use the benchmark runner to evaluate LTO gains (on clang and gcc) and performance differences between clang and gcc. Role of these tests is mostly checking that LTO and CMAKE_LLVM_PATH options keep working over time.

Current invocation of regression script involves lots of copy pasting, probably logic should be refactored, but at first I though this was clearer.