@carlopi 2023-07-25
There is still an unconnected failure in Linux job here: https://github.com/duckdb/duckdb/actions/runs/5659062468/job/15331786339?pr=8357#step:7:2949.

I run a few experiments as part of the tests, using geometric mean of current regression tests.
Results are:

| image | base | alternative | micro | tpch | tpcds | h2oai | imdb |
|--|---|--|---|--|--|---|---|
| macos-latest | clang 14 | clang 14, LTO=full | -17% | -3% | -2% | -5% | -5% |
| macos-latest | clang 14 | clang 16 | -15% | same | -2% | same | same |
| ubuntu-latest | gcc 11.3 | clang 14 | -7% | +2% | -4% | -2% | -3% |
| ubuntu-latest | gcc 11.3 | gcc 11.3, LTO=full | -1% | -2% | -1% | -3% | -1% |

Unsure what should be read here, something like:
clang marginally better than gcc, newer compiler version marginally better than older compiler versions, gcc LTO marginally better than regular gcc, clang LTO visibly better than regular clang.
Takeaway is that probably best combination is recent clang with LTO enabled, but hard to put actual numbers, and to be balanced with bringing in additional dependencies or existing restrictions.