@carlopi 2023-07-26
Thanks @samansmink and @Mytherin for the feedback.

I guess the hard choice that is what to make of all this / what to turn on and on what condition, and that would probably require some added considerations. On this PR I took the easy road of just providing options (and avoiding to have to re-discover the set of changes that were needed again in the future) without making real choices.

This PR is having another round of CI since I moved the benchmarks to a separate workflow (to be run only via workflow dispatch or on changes to the workflow itself), but on my side is ready to be merged.

---
If/when we want to experiment, I would consider probably easier to do so on OSX-based workflows, given that there clang it's already the default + we have an easier time testing them, and potentially moving from clang 14 to brew installed clang 16 and enabling LTO can bring gains with lower risks.