@Mytherin 2023-08-07
Thanks - we can merge this and leave the actual choice of whether/where we want to enable LTO for a later date.